requirements.txt
flask==2.2.3
flask-pymongo==2.3.0
python-dotenv==1.0.0
requests==2.28.2
Werkzeug==2.2.3
google-api-python-client==2.86.0
spacy==3.5.3
nltk==3.8.1
